requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
gunicorn==20.1.0
numpy>=1.24.0 